ログイン機能を実装

+++ app/Http/Controllers/Tweet/DeleteController.php
@@ -4,16 +4,22 @@ namespace App\Http\Controllers\Tweet;
 
 use App\Http\Controllers\Controller;
 use App\Models\Tweet;
+use App\Services\TweetService;
 use Illuminate\Http\Request;
+use Symfony\Component\HttpKernel\Exception\AccessDeniedHttpException;
 
 class DeleteController extends Controller
 {
     /**
      * Handle the incoming request.
      */
-    public function __invoke(Request $request)
+    public function __invoke(Request $request, TweetService $tweetService)
     {
         $tweetId = $request->route('tweetId');
+        if (!$tweetService->isOwnTweet($request->user()->id, $tweetId)) {
+            throw new AccessDeniedHttpException();
+        }
+        
         $tweet = Tweet::where('id', $tweetId)->firstOrFail();
         $tweet->delete();
